Refactored Switch with Typescript
