+++ src/elements/Switch/index.tsx
@@ -1,0 +1,3 @@
+import Switch from './Switch';
+
+export default Switch;
